test(flags): Instantiate needed filters from the test class

+++ tests/phpunit/tests/test-flags.php
@@ -13,26 +13,28 @@ class Flags_Test extends PLL_UnitTestCase {
 		self::create_language( 'en_US' );
 		self::create_language( 'fr_FR' );
 
-		@mkdir( WP_CONTENT_DIR . '/polylang' );
-		copy( dirname( __FILE__ ) . '/../data/fr_FR.png', WP_CONTENT_DIR . '/polylang/fr_FR.png' );
-	}
-
-	static function wpTearDownAfterClass() {
-		parent::wpTearDownAfterClass();
-
-		unlink( WP_CONTENT_DIR . '/polylang/fr_FR.png' );
-		rmdir( WP_CONTENT_DIR . '/polylang' );
+		$wp_filter['pll_languages_list']->remove_all_filters();
+		$wp_filter['pll_after_languages_cache']->remove_all_filters();
 	}
 
 	function setUp() {
 		parent::setUp();
 
 		$options       = array_merge( PLL_Install::get_default_options(), array( 'default_lang' => 'en_US' ) );
-		$model         = new PLL_Model( $options );
-		$links_model   = new PLL_Links_Default( $model );
-		$this->pll_env = new PLL_Frontend( $links_model );
-		$this->pll_env->init();
-		$this->pll_env->model->cache->clean();
+		$this->model         = new PLL_Model( $options );
+		$links_model = new PLL_Links_Default( $model ); // Registers the 'pll_languages_list' and 'pll_after_languages_cache' filters.
+	}
+
+	public function tearDown() {
+		if ( file_exists( WP_CONTENT_DIR . '/polylang/fr_FR.png' ) ) {
+			unlink( WP_CONTENT_DIR . '/polylang/fr_FR.png' );
+			rmdir( WP_CONTENT_DIR . '/polylang' );
+		}
+
+		if ( isset( $_SERVER['HTTPS'] ) ) {
+			unset( $_SERVER['HTTPS'] );
+		}
+		parent::tearDown();
 	}
 
 	function test_default_flag() {
@@ -42,6 +44,9 @@ class Flags_Test extends PLL_UnitTestCase {
 	}
 
 	function test_custom_flag() {
+		@mkdir( WP_CONTENT_DIR . '/polylang' );
+		copy( dirname( __FILE__ ) . '/../data/fr_FR.png', WP_CONTENT_DIR . '/polylang/fr_FR.png' );
+
 		$lang = self::$model->get_language( 'fr' );
 		$this->assertEquals( content_url( '/polylang/fr_FR.png' ), $lang->get_display_flag_url() );
 		$this->assertEquals( '<img src="/wp-content/polylang/fr_FR.png" alt="Français" />', $lang->get_display_flag() );
@@ -55,17 +60,15 @@ class Flags_Test extends PLL_UnitTestCase {
 
 		$lang = self::$model->get_language( 'en' );
 		$this->assertContains( 'https', $lang->get_display_flag_url() );
-
-		unset( $_SERVER['HTTPS'] );
 	}
 
 	function test_custom_flag_ssl() {
 		$_SERVER['HTTPS'] = 'on';
+		@mkdir( WP_CONTENT_DIR . '/polylang' );
+		copy( dirname( __FILE__ ) . '/../data/fr_FR.png', WP_CONTENT_DIR . '/polylang/fr_FR.png' );
 
 		$lang = self::$model->get_language( 'fr' );
 		$this->assertEquals( content_url( '/polylang/fr_FR.png' ), $lang->get_display_flag_url() );
 		$this->assertContains( 'https', $lang->get_display_flag_url() );
-
-		unset( $_SERVER['HTTPS'] );
 	}
 }
